+++ glfw/wl_client_side_decorations.h
@@ -14,3 +14,4 @@ void free_csd_surfaces(_GLFWwindow *window);
 void change_csd_title(_GLFWwindow *window);
 bool ensure_csd_resources(_GLFWwindow *window);
 void set_csd_window_geometry(_GLFWwindow *window, int32_t *width, int32_t *height);
+void set_titlebar_color(_GLFWwindow *window, uint32_t color, bool use_system_color);