+++ kitty/config_data.py
@@ -1260,6 +1260,14 @@ def macos_titlebar_color(x: str) -> int:
     return (color_as_int(to_color(x)) << 8) | 2
 
 
+o('wayland_titlebar_color', 'system', option_type=macos_titlebar_color, long_text=_('''
+Change the color of the kitty window's titlebar on Wayland systems with client side window decorations such as GNOME.
+A value of :code:`system` means to use the default system color,
+a value of :code:`background` means to use the background color
+of the currently active window and finally you can use
+an arbitrary color, such as :code:`#12af59` or :code:`red`.
+'''))
+
 o('macos_titlebar_color', 'system', option_type=macos_titlebar_color, long_text=_('''
 Change the color of the kitty window's titlebar on macOS. A value of :code:`system`
 means to use the default system color, a value of :code:`background` means to use